% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_covered.R
\name{get_covered}
\alias{get_covered}
\title{Get locations within a specified distance of another set of locations}
\usage{
get_covered(install_at, to_cover, cover_distance = units::set_units(25, "km"))
}
\arguments{
\item{install_at}{A spatial data frame of installation points}

\item{to_cover}{A spatial data frame of points to cover}

\item{cover_distance}{The distance each installation point covers}
}
\value{
A tibble data frame with two columns:
\itemize{
\item \code{to_cover_id}: the row indices of \code{to_cover} that are within the \code{cover_distance} of \code{install_at}
\item \code{install_at_id}: the row indices of \code{install_at} that cover \code{to_cover_id}
A \code{to_cover} location may be covered by multiple (or no) \code{install_at} locations.
}
}
\description{
Checks whether each location in \code{to_cover} is within \code{cover_distance} of each location in \code{install_at}.
Leverages the fast \code{\link[sf:geos_binary_pred]{sf::st_is_within_distance()}} function to perform the check, so distances are not provided.
See \code{\link[=mark_nearest_location]{mark_nearest_location()}} if the nearest distance is needed as well.
}
