% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_nearest_location.R
\name{mark_nearest_location}
\alias{mark_nearest_location}
\title{Add identifiers marking the nearest locations and their distance to each point.}
\usage{
mark_nearest_location(
  from,
  to,
  within = units::set_units(Inf, "km"),
  from_id_col = ".id",
  to_id_col = ".id"
)
}
\arguments{
\item{from}{\code{sf} data frame containing the points to mark}

\item{to}{\code{sf} data frame containing the points to mark nearest to}

\item{within}{distance to mark nearest points within. Either a \code{units} object (see \link[units:units]{units::set_units}), or a \code{numeric} in km}

\item{from_id_col, to_id_col}{column name in \code{from}/\code{to} containing unique row identifiers. If not present will be included (using \link[dplyr:row_number]{dplyr::row_number}) and removed within the function.}
}
\value{
data frame with the nearest location in \code{to} to each location in \code{from} marked.
Columns \code{.nearest_to_id}, \code{.nearest_to_distance}, and \code{.has_nearby} will be added,
and if \code{to} is grouped, those will be repeated for each grouping as well.
If \code{from} is grouped the results will not be different from ungrouped, but the grouping will be maintained.
}
\description{
This function marks the nearest location (id and distance) from a set of points to another set of points.
For locations within \code{within}, a flag is raised indicating that site has a location nearby.
If \code{to} is a grouped dataframe (see \link[dplyr:group_by]{dplyr::group_by}), the nearest location of each grouping is marked, as well as the normal overall ungrouped columns.
}
\details{
This function relies on unique row identifier columns for both \code{from} and \code{to}, proovided via the \code{from_id_col} and \code{to_id_col} arguments.
If no such column is present, a row identifier column is added with default name \code{".id"}, and will be removed before returning the output.
}
