% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_coverage.R
\name{optimize_coverage}
\alias{optimize_coverage}
\title{Optimize coverage of a set of locations by a set of installations.}
\usage{
optimize_coverage(
  install_at,
  to_cover,
  existing_locations = NULL,
  cover_distance = units::set_units(25, "km"),
  weight_columns = c(".weight", ".weight")
)
}
\arguments{
\item{install_at}{An \code{sf} data frame containing the locations of the installations.}

\item{to_cover}{An \code{sf} data frame containing the locations of the places to cover with installations.}

\item{existing_locations}{A data frame containing the locations of existing installations.
\code{to_cover} will be filtered to exclude points that are already covered by \code{existing_locations}.}

\item{cover_distance}{The distance from an installation that a point is considered to be covered.
Expected to be a \code{units} object, otherwise assumed to be in km.
Defaults to 25 km.}

\item{weight_columns}{A character vector containing the names of the columns in \code{install_at} and \code{to_cover}
that should be used as weights when calculating the coverage of each installation.
A weight of 2 is akin to double coverage from/for that location.
If no columns match the names in \code{weight_columns}, equal weighting is assumed.
If only one column is provided, it will be used for both \code{install_at} and \code{to_cover}.
Defaults to c(".weight", ".weight").}
}
\value{
A data frame containing the subset of installations that cover the most points with the least number of installations, weighted by \code{weight_columns}.
}
\description{
This function takes a set of points to cover and a set of installations
and returns the subset of installations that cover the most points within a set distance
with the least number of installations. It allows for weighting of
points to cover and installations based on columns in the data frames.
}
