% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coverage.R
\name{plot_coverage}
\alias{plot_coverage}
\title{Plot the coverage of a set of installations over a set of points.}
\usage{
plot_coverage(
  install_at,
  to_cover,
  existing_locations = NULL,
  optimized_locations = NULL,
  cover_distance = units::set_units(25, "km"),
  weight_columns = c(".weight", ".weight"),
  background_map = NULL,
  in_canada = FALSE,
  select_prov_terrs = "all",
  colours = list(to_cover = "#FC4E07", existing = "#00AFBB", proposed = "#E7B800", others
    = "black"),
  fill_labels = list(to_cover = "To Cover", existing = "Existing", proposed = "Proposed",
    others = "Others"),
  fill_opacity = 0.7,
  point_shape = 21,
  stroke_colour = "black"
)
}
\arguments{
\item{install_at}{An \code{sf} data frame containing locations of potential installations.}

\item{to_cover}{An \code{sf} data frame containing the locations desired to be covered by \code{install_at} within \code{cover_distance}.}

\item{existing_locations}{An \code{sf} data frame containing the locations of existing installations.
Defaults to \code{NULL} - no existing locations will be plotted.}

\item{optimized_locations}{An \code{sf} data frame containing the optimized locations for installation.
Defaults to \code{NULL} - no optimized locations will be plotted.}

\item{cover_distance}{The distance from an installation that a location is considered to be covered.
Expected to be a \code{units} object, otherwise a \code{numeric} which is assumed to be in km.
Defaults to 25 km.}

\item{weight_columns}{A character vector containing the names of the columns in \code{install_at} and \code{to_cover}
that was used as weights when optimizing the coverage of each potential installation.}

\item{background_map}{A \code{ggplot} object containing the base map of the plot.
If provided, the coverage will be plotted on top of this map.
Defaults to \code{NULL} - a blank plot will be used.}

\item{in_canada}{A logical indicating whether the coverage should be plotted over a map of Canada.
If \code{TRUE}, a background map of Canada will be generated if \code{background_map} is \code{NULL}, and the plot will be projected using a Canada-specific projection.
Defaults to \code{FALSE}.}

\item{select_prov_terrs}{A character vector containing the names (abbreviations or EN/FR names) of the provinces and territories to include in the map of Canada.
Ignored if \code{in_canada} is \code{FALSE}.
Defaults to \code{"all"} - meaning a bounding box will be calulated from the extent of the provided data for the map instead of filtering by province/territory.}

\item{colours}{A list containing the fill colours for coverage and installation locations.
Must be length 4 and have names \code{to_cover}, \code{existing}, \code{proposed}, and \code{others}.
Defaults to \code{list(to_cover = "#FC4E07", existing = "#00AFBB", proposed = "#E7B800", others = "black")}.}

\item{fill_labels}{A list containing the labels to display in the legend got the coloured layers.
Must be length 4 and have names \code{to_cover}, \code{existing}, \code{proposed}, and \code{others}.
Defaults to \code{list(to_cover = "To Cover", existing = "Existing", proposed = "Proposed", others = "Others")}.}

\item{fill_opacity}{A numeric between 0 and 1 indicating the opacity of the coverage layers.
Defaults to 0.7.}

\item{point_shape}{The shape to use for the non-prioritized installation locations.
Defaults to 21 (a filled circle).}

\item{stroke_colour}{The colour to use for strokes throughout the plot.}
}
\value{
A \code{ggplot} object containing the coverage of the installations over the points.
}
\description{
Plot the coverage of a set of installations over a set of points.
}
