% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritize_installations.R
\name{prioritize_installations}
\alias{prioritize_installations}
\title{Prioritize installations by their coverage.}
\usage{
prioritize_installations(
  install_at,
  to_cover,
  cover_distance = units::set_units(25, "km"),
  weight_columns = c(".weight", ".weight"),
  suffix = ""
)
}
\arguments{
\item{install_at}{An \code{sf} data frame containing the locations of the to consider for installation.}

\item{to_cover}{An \code{sf} data frame containing the locations desired to be covered by \code{install_at} within \code{cover_distance}.
If grouped, the coverage will be calculated for each group and the sum of those will be used to prioritize
(effectively this adds additional information on coverage of each group without changing the prioritization).}

\item{cover_distance}{The distance from an installation that a location is considered to be covered.
Expected to be a \code{units} object, otherwise a \code{numeric} which is assumed to be in km.
Defaults to 25 km.}

\item{weight_columns}{A character vector containing the names of the columns in \code{install_at} and \code{to_cover}
that should be used as weights when prioritzing the coverage of each potential installation.
Weights will be summed for all \code{to_cover} covered by each \code{install_at}, and multiplied by the corresponding weight for \code{install_at}.
If a single weight column is provided, it will be used for both \code{install_at} and \code{to_cover}.
If not found in \code{install_at} or \code{to_cover}, a default of weight 1 will be used.
Defaults to \code{c(".weight", ".weight")}, equivalent to \code{".weight"}.}

\item{suffix}{A character string to append to the end of the new columns created (i.e. \code{"newly_covered{suffix}"} and \code{"priority{suffix}"}).
Defaults to \code{""}.}
}
\value{
An \code{sf} data frame ranking the installations weighted by \code{weight_columns} with added columns for \code{"newly_covered{suffix}"} and \code{"priority{suffix}"} (as well as additional pairs of columns for each group in \code{to_cover} if grouped).
The \code{"newly_covered{suffix}"} column contains the sum product of weights for new locations to cover by each installation if installed in order of \code{"priority{suffix}"}.
}
\description{
Given a set of points to cover (\code{to_cover}) and a set of points to install at (\code{install_at}), prioritizes the points to install at by their coverage of points to cover.
The prioritization is based on the weighted sum of the coverage of each point to install at.
The weights are determined by the columns specified in \code{weight_columns}.
The coverage of each point to install at is determined by the distance \code{cover_distance}.
}
